LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-from-data-url-expected.txt

CONSOLE MESSAGE: line 1: TypeError: Undefined value
This tests that a data: URL loaded in an iframe inside another data: URL loaded iframe doesn't have access to its parent, the first data: URL loaded iframe.




--------
Frame: 'aFrame'
--------
PASS: Cross frame access from a data: URL was denied.

Inner iframe.



--------
Frame: 'innerFrame'
--------
Inner-inner iframe.



--------
Frame: '<!--framePath //aFrame/innerFrame/<!--frame0-->-->'
--------
